// nodejs/external-user-service/user-service/src/domain.types/users/google.oauth.types.ts
export interface GoogleAccessTokenModel {
    access_token: string;
    expires_in: number;
    refresh_token?: string;
    scope: string;
    token_type: string;
    id_token?: string;
}

export interface GoogleUser {
    id: string;
    email: string;
    verified_email: boolean;
    name: string;
    given_name: string;
    family_name: string;
    picture?: string;
    locale?: string;
}

export interface GoogleOAuthCallbackParams {
    code: string;
    state?: string;
    scope?: string;
}